+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.6] - 2026-07-06
+
 ### Fixed
 
 - Timestamps rendering in the future (e.g. Log Events "last seen: in 1h") on